Additions to Activities to allow resource management
This is by no means finished. There are plenty of things to do - but I send it out now so that you can have a look at it as it is to see what my plans are.

Note: The activities management functions are not yet called. So there should be no impact on anything running yet. 

Please have a look! If you generally agree I will try to get it in before I go on vacation, otherwise it will have to wait until I get back. I might even work tomorrow instead of taking off as planned if needed.

BUG=388085
TEST=not yet done
TBR=rsesek

Review URL: https://codereview.chromium.org/358003002