Mojo surfaces service and example app
This provides an implementation of the mojo surfaces service and a
sample with two mojo applications composing each other via surfaces. The
viewport / display hookup is a bit of a hack in this patch as is the
complete lack of scheduling logic, but it's enough to get started.

Review URL: https://codereview.chromium.org/361123002