[telemetry] Reduce browser launches in unit tests.
Instead of launching the browser once per test, we can launch it once per TestCase class. This has increased risk of side effects, but we judge the reduction in cycle time to be worth the tradeoff.

On my Linux workstation, this decreases telemetry_unittests cycle time from 4:22 to 3:07, or a 30% reduction. The difference may be even greater on Android.


BUG=337829, 379378, 388256
TEST=tools/telemetry/run_tests

Review URL: https://codereview.chromium.org/383323002