Place OEM folder for new users after Web Store
It turns out that on some OEM devices the Files app is placed near the
end, so placing the OEM folder before the last non-removable app does
not have the desired result.

Instead, we place it after the WebStore, and assume that a pre-installed
app (e.g. Search) exists so that the app location is stable.

This was tested succesfully on an official build on one of the failing
devices. (I also confirmed that all of the existing defailt_app_order
files have the WebStore followed by Search early in the list).

BUG=chrome-os-partner:29764

Review URL: https://codereview.chromium.org/396843003