Remove X11/Xlib.h include from selection_requestor.h
Removing the include of Xlib.h makes selection_requestor.h more easily
testable. In particular, the definition of None in X.h conflicts with
that in gtest-type-util.h

This CL also adds XAtom to x11_types.h so that it does not need to be forward declared everywhere.

BUG=None
TEST=None
R=derat,danakj
TBR=dcheng (for trivial change to clipboard_aurax11.cc)

Review URL: https://codereview.chromium.org/392153002