Ensure pdf plugin sends DidStartLoading() consistently
There are situations where the pdf plugin will initialize itself in such a way
that it never notifies content/renderer/ that it has started. In so doing, it
also fails to send a stop message, since it tries to ensure it doesn't send
mismatched start/stop messages.

Print preview decides when to display the preview based on when loading
stops. Since this only considers the page's content and not the pdf plugin's
initialization, there is a race condition where we may try to preview before
the plugin is ready. In this cases, we show "Loading preivew..." and never load
the actual preview.

There is a hack in place in print_preview_web_helper.cc to delay the process of
showing the preview to maximize the chance that the pdf plugin will be ready.
This change should allow us to rip that out.

BUG=376969

Review URL: https://codereview.chromium.org/397713005