Set callback interface on WebCDMSessionImpl after creation.
In order to allow MediaKeySession to be created only when resolving
the promise, add the ability to set the client callback interface on
WebContentDecryptionModuleSessionImpl after the object is created.

This will be used in a subsequent CL to update blink. Requires a third
change to remove the old methods once blink is updated.

BUG=358271
TEST=EME layouts tests pass

Review URL: https://codereview.chromium.org/380953002