Update insets on change to window bounds.
In overscroll mode, any window with web contents that overlaps the virtual keyboard has insets in order to facilitate scrolling to the bottom of the document.  These insets are updated whenever the keyboard is shown or hidden.  This patch adds another update in the event that a window moves while the keyboard is visible.  This case can occur after a screen rotation while the keyboard is shown.


BUG=391337

Review URL: https://codereview.chromium.org/368323002