Fix extension scrollbar regression bug, and add a regression test to catch similar issues in the future.
This changes the extension_install_dialog_view permission strings layout to have a fixed width for the bullet, centered, and subtract that width from the width used to size the permission string label.

This also adds two browser tests: one to test that a scrollbar _does_ show for a very long set of permissions, and one for the specific regression case from the bug.

BUG=385570

Review URL: https://codereview.chromium.org/363523002