Check the actual size of canvas to update.
Previously the gallery checks the sequence number to find out whether the update
for canvas size is needed or not.  The sequence number is incremented when
variables that affect canvas size are updated. But sometimes, the
canvas size goes to be out of sync.

BUG=245926
TEST=manually

Review URL: https://codereview.chromium.org/399383002