Cast: Refactor RTCP handling
OBJECTIVES

This change is to refactor RTCP handling in media/cast such that RTCP
message parsing is done in CastTransportSender.

This side effect of this refactoring is much cleaner code: RTCP handling
code are grouped under one module. VideoSender and AudioSender do not
handle packets directly.

This also affect the architectutre in Chrome that RTCP message parsing
is now done in the browser process. A parsed RTCP message is passed to
the renderer instead of raw packets.

REATIONALE

RTCP is used as a feedback mechanism in Cast Streaming. It used to be
parsed and handled by VideoSender and AudioSender in the render process.
This was very ineffective because packetization and sending is done in
the browser process. This created inefficiencies in packet
retransmission. It also made improvement to the transport protocol /
algorithm much more difficult. A side effect is very messy code.

DETAILS

Rtcp - This class is responsibile for maintaining a bi-directional RTCP
session. It is now owned by CastTransportSender.

CastTransportSender - It now performs packetization and also parsing of
incoming RTCP packets. RTCP packets are small UDP packets for signaling
only. Eventually retransmission can also be moved to this class. It will
handle all packet level send transport.

AudioSender / VideoSender - They don't handle packet directly any more.
Which means render process doesn't see raw packets.

FrameSender - Base class for AudioSender and VideoSender to share code
for RTCP message handling. Eventually AudioSender and VideoSender will
merge here but it is not the point of this change.

RtcpBuilder - Removed and merged with RtcpSender.

Net deleted 400 lines of code and simulation shows the same results.

BUG=393042

Review URL: https://codereview.chromium.org/387933005