Enables permission bubbles to remain visible during fullscreen on Mac.
Adds IsVisible() function to PermissionBubbleView so that the browser window
knows if it is currently being shown.  It's added to the base class, rather than
just the cocoa implementation, because similar functionality will likely be
used for views in the future.

Adds alwaysShowDropdown property to the PresentationModeController.
When set to YES, it disables hiding of the dropdown - toolbar and tab strip.

Also adds a hack to ensure the UI is drawn correctly when entering fullscreen
mode.  Specifically, for the duration of the animation, the top-level view for
the browser window has setWantsLayer set to YES.  Without this, tabStripView
is not rendered into the destination frame for the animation.  (It is a sibling of
the content view, not a subview, and I believe that the non-layer-based
render path bypasses it when the animation begins.)  I could not come up
with any other mechanism to fix this, and without it, fullscreen looks quite
bad - the toolbar displays properly, but the tab strip is missing.

BUG=384260

Review URL: https://codereview.chromium.org/390503003