Do not assign launch ordinals to ephemeral apps
Ephemeral apps are not shown in the app launcher or NTP, but they are
assigned ordinals, which results in unfilled NTP pages. This patch
does not give ephemeral apps launch ordinals until they are promoted
to fully installed apps.

AppSorting::MarkExtensionAsHidden() was refactored to
AppSorting::SetExtensionVisible() to allow the visibility of the
extension in the NTP to change.

BUG=394192
TEST=browser_tests

Review URL: https://codereview.chromium.org/397903002