Add app icon masp to notification API
Add an app icon mask (16*16, or 32*32 for retina display) on the bottom right corner of notifications to indicate source of the notification. The app icon is displayed in a way that takes only the alpha channel of the image passed in, masks it with white foreground and grey background.

BUG=284592

Review URL: https://codereview.chromium.org/308053004