[Ozone-DRI] Convert HardwareDisplayController to use scanout buffers
This change allows scanout buffers to be kept alive until the last of
SurfaceOzone{Canvas,EGL} and HardwareDisplayController are destroyed. For
HardwareDisplayController this means that buffers will be kept around
until the CRTC is properly restored. For SurfaceOzoneEGL this means that
on monitor disconnects (HDC is destroyed) the buffers are kept alive for
any remaining EGL operations still happening.

BUG=none
NOTRY=true

Review URL: https://codereview.chromium.org/393233005