Remaining bits to get views_examples_with_content_exe to work on Mac
The example window currently doesn't activate itself properly because it
lacks an activation policy and a menubar.

This CL sets the activation policy, adds a menubar and defers window
creation until the applicationDidFinishLoading notification. Creating a
window before this can activate and receive key events, but the menubar
will not switch properly.

With this, views_examples_with_content_exe compiles and runs nicely on a
Mac build with `GYP_DEFINES=toolkit-views=1`. Note the `WebView` example
will crash for now because views::NativeViewHost is not yet implemented
for mac.

BUG=395507

Review URL: https://codereview.chromium.org/309483009