Move the shared timer to be dynamically allocated.
The shared timer must be deleted on the same thread that it was executing on. When the WebThread is being destroyed, it is done from a different thread, so the shared timer must be deleted prior to that. This gives an API for the thread cleanup code to do so.

BUG=301515
R=jochen@chromium.org

Review URL: https://codereview.chromium.org/416453002