Minimize HomeCard if the user opens something in the home card.
This is applied to:
- open to the search result
- click on app icon

I am not sure this is the right approach. We may want to introduce
an ActivityObserver or something, and minimize the home card
by itself. However, activity may be created for various reason
but minimizing should work upon the user's action on the
home card, so this approach isn't bad.

BUG=None
R=oshima@chromium.org, sadrul@chromium.org
TBR=sky@chromium.org
TEST=manually

Review URL: https://codereview.chromium.org/408623003